[#58] 메뉴 옵션 기능 추가 및 수정 (#92)
* MenuOption Controller 기능 추가 및 수정

* 불 필요 파일 삭제

* MenuOption 서비스 기능 추가 및 수정

* MenuOption 엔티티 및 mapper 추가 및 수정

* OptionRequest 수정 toParam추가

* OptionController 수정. 서비스 호출 param 변경

* Update MenuOptionService.java

* Update MenuOptionController.java

* MenuOptionServiceImpl 수정

* Update MenuOptionServiceImpl.java

+++ src/main/java/com/flab/foodeats/application/menu/adapter/MenuOptionServiceImpl.java
@@ -1,46 +1,71 @@
 package com.flab.foodeats.application.menu.adapter;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import com.flab.foodeats.application.menu.OptionTarget;
+import com.flab.foodeats.api.menu.OptionRequest;
 import com.flab.foodeats.application.menu.port.MenuOptionService;
+import org.springframework.transaction.annotation.Transactional;
+import com.flab.foodeats.domain.menu.MenuOption;
+import com.flab.foodeats.infra.menu.MenuOptionMapper;
+import com.flab.foodeats.infra.shop.ShopMapper;
 
 @Service
+@Transactional
 public class MenuOptionServiceImpl implements MenuOptionService {
 
 	private final MenuOptionMapper menuOptionMapper;
+	private final UserMapper userMapper;
 
-	public MenuOptionServiceImpl(MenuOptionMapper menuOptionMapper) {
+	public MenuOptionServiceImpl(MenuOptionMapper menuOptionMapper, UserMapper userMapper) {
 		this.menuOptionMapper = menuOptionMapper;
+		this.userMapper = userMapper;
 	}
 
 	@Override
-	public void registerMenuOption(int menuId, OptionTarget target) {
-		List<MenuOption> menuOptions = target.getMenuOptions();
-		menuOptions.stream()
-			.forEach(option ->
-				menuOptionMapper.registerOption(menuId, option.getMenuOptionName(), option.getMenuOptionPrice()));
+	public void registerMenuOption(Long shopId, int menuId, List<OptionRequest> optionRequests, String userId) {
+		String requestedOwnerId = userMapper.findMerchantByShopId(shopId).getUserId();
+		checkIsAuthorizedUser(requestedOwnerId, authInfo.getUserId());
+
+		List<MenuOption> menuOptions = menuOptionMapper.searchMenuOption(menuId);
+		checkIsExistOption(optionRequests, menuOptions);
+
+		optionRequests.stream()
+			.forEach(optionRequest -> menuOptionMapper.registerOption(optionRequest.toEntity(menuId)));
 	}
 
 	@Override
-	public List<MenuOption> searchMenuOption(int menuId) {
-		return menuOptionMapper.searchMenuOption(menuId);
+	public void updateMenuOption(Long shopId, OptionTarget optionTarget, AuthInfo authInfo) {
+		String requestedOwnerId = userMapper.findMerchantByShopId(shopId).getUserId();
+		checkIsAuthorizedUser(requestedOwnerId, authInfo.getUserId());
+		
+		MenuOption option = menuOptionMapper.searchMenuOptionByOptionId(optionTarget.getMenuOptionId());
+		option.updateOption(optionTarget.getMenuOptionName(), optionTarget.getMenuOptionPrice());
+		menuOptionMapper.updateMenuOption(option);
 	}
 
 	@Override
-	public void updateMenuOption(int menuOptionId, OptionTarget target) {
-		List<MenuOption> menuOptions = target.getMenuOptions();
-		for (MenuOption menuOption : menuOptions) {
-			String menuOptionName = menuOption.getMenuOptionName();
-			int menuOptionPrice = menuOption.getMenuOptionPrice();
-			menuOptionMapper.updateMenuOption(menuOptionId, menuOptionName, menuOptionPrice);
+	public void deleteMenuOption(Long shopId, int menuOptionId, String userId) {
+		checkIsAuthorizedUser(shopId, userId);
+		menuOptionMapper.deleteMenuOption(menuOptionId);
+	}
+
+	private void checkIsAuthorizedUser(String requestedId, String userId) {
+		if (!requestedId.equals(userId)) {
+			throw new UnauthorizedException();
 		}
 	}
 
-	@Override
-	public void deleteMenuOption(int menuOptionId) {
-		menuOptionMapper.deleteMenuOption(menuOptionId);
+	private void checkIsExistOption(List<OptionRequest> optionRequests, List<MenuOption> menuOptions) {
+		for (OptionRequest optionRequest : optionRequests) {
+			Optional<MenuOption> existOption = menuOptions.stream()
+				.filter(m -> m.getMenuOptionName().equals(optionRequest.getMenuOptionName()))
+				.findAny();
+			if (existOption.isPresent()) {
+				throw new IllegalArgumentException(MenuCode.ALEADY_EXIST_MENU.getMessage());
+			}
+		}
 	}
 }
